Move ProfileSettings to own file
